added some methods to creature, player hitpoints now double

+++ AdventureGame/Player.java
@@ -1,8 +1,8 @@
 public class Player {
 
     Room currentRoom;
-    int hitpoints;
-    final int maxHitpoints = 100;
+    double hitpoints;
+    final double maxHitpoints = 100;
     AdventureGame adventureGame;
 
     public Player(Room startRoom, AdventureGame adventureGame) {
@@ -12,8 +12,8 @@ public class Player {
     }
 
     public Room getRoom() { return this.currentRoom; }
-    public int getHitpoints() { return this.hitpoints; }
+    public double getHitpoints() { return this.hitpoints; }
     public void moveToRoom(Room room) { this.currentRoom = room; }
-    public void setHitpoints(int hitpoints) { this.hitpoints = hitpoints; }
+    public void setHitpoints(double hitpoints) { this.hitpoints = hitpoints; }
 
 }
